requirements.txt
appnope==0.1.2
backcall==0.2.0
cycler==0.10.0
decorator==5.0.9
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.20.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==22.1.0
six==1.16.0
snap-stanford==6.0.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
